% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/googleNgrams-package.R
\docType{package}
\name{googleNgrams-package}
\alias{googleNgrams}
\alias{googleNgrams-package}
\title{googleNgrams: What the Package Does (Title Case)}
\description{
More about what it does (maybe more than one line) Use four spaces when indenting paragraphs within the Description.
}
\keyword{internal}
