% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gram_build_df.R
\name{gram_build_dtm}
\alias{gram_build_dtm}
\title{Convert ngram output to dtm}
\usage{
gram_build_dtm(x)
}
\arguments{
\item{x}{A data frame}
}
\value{
A data frame
}
\description{
Convert ngram output to dtm
}
