% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gram_get_unzip.R
\name{gram_get_unzip}
\alias{gram_get_unzip}
\title{Download, unzip ngram file as csv}
\usage{
gram_get_unzip(url, dir = tempdir())
}
\arguments{
\item{url}{A character string}

\item{dir}{A character string}
}
\value{
A list
}
\description{
Download, unzip ngram file as csv
}
