% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gram_sample_ngram.R
\name{gram_sample_ngram}
\alias{gram_sample_ngram}
\title{Sample, filter ngram dtm}
\usage{
gram_sample_ngram(
  x,
  start_date,
  end_date,
  seed = 99,
  generation = 25,
  remove_nonalpha = T,
  tolower = T,
  sample = 5e+05
)
}
\arguments{
\item{x}{A data frame}

\item{start_date}{An integer}

\item{end_date}{An integer}

\item{seed}{An integer}

\item{generation}{An integer}

\item{remove_nonalpha}{A boolean}

\item{tolower}{A boolean}

\item{sample}{An integer}
}
\value{
A list
}
\description{
Sample, filter ngram dtm
}
